+++ game/rook.py
@@ -6,7 +6,6 @@ current_file_path = os.path.abspath(__file__)  # Obtener la ruta absoluta del ar
 project_root = os.path.dirname(os.path.dirname(current_file_path))  # Obtener el directorio raíz del proyecto
 sys.path.append(project_root)  # Agregar el directorio raíz al sys.path
 
-
 from game.piece import Piece
 
 class Rook(Piece):
@@ -15,7 +14,7 @@ class Rook(Piece):
         self.color = color
 
     def __str__(self):
-        return '♜' if self.color == "WHITE" else '♖'  # Devuelve '♜' si el color de la torre es blanca, sino devuelve '♖'
+        return '♜' if self.color == "WHITE" else '♖'  # Símbolos para la torre
 
     def move(self, start_pos, end_pos, board):
         start_row, start_col = start_pos
@@ -31,6 +30,7 @@ class Rook(Piece):
 
         current_row, current_col = start_row + step_row, start_col + step_col
         
+        # Iterar a través de las casillas entre la posición de inicio y la de destino
         while (current_row, current_col) != (end_row, end_col):
             if board[current_row][current_col] is not None:
                 return False
@@ -39,5 +39,4 @@ class Rook(Piece):
 
         # Verificar si la casilla de destino está vacía o contiene una pieza del oponente
         destination_piece = board[end_row][end_col]
-        is_empty_or_enemy = destination_piece is None or destination_piece.color != self.color
-        return is_empty_or_enemy
+        return destination_piece is None or destination_piece.color != self.color